# docs/swagger.yaml
basePath: /v1
definitions:
  api.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  domain.GetUserByIDResponse:
    properties:
      user:
        $ref: '#/definitions/domain.User'
    type: object
  domain.SignInRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  domain.SignInResponse:
    properties:
      access_token:
        type: string
      expires_at:
        type: string
      user:
        $ref: '#/definitions/domain.User'
    type: object
  domain.SignUpRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
  domain.SignUpResponse:
    type: object
  domain.User:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
      verified:
        type: boolean
    type: object
info:
  contact: {}
  description: The Nexthook API.
  title: Nexthook API
  version: "1.0"
paths:
  /auth/signin:
    post:
      consumes:
      - application/json
      description: sign in a user
      parameters:
      - description: Signin Object
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.SignInRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.SignInResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: User Sign In
      tags:
      - auth
  /auth/signup:
    post:
      consumes:
      - application/json
      description: sign up a new user
      parameters:
      - description: Signup Object
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.SignUpRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.SignUpResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: User Sign Up
      tags:
      - auth
  /users/{id}:
    get:
      consumes:
      - application/json
      description: get a user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.GetUserByIDResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - user
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"